pyproject.toml
[project]
name = "flappy-bird"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neat-python>=0.92",
    "pygame>=2.6.1",
]
